fixed bug in jarRoutes
