+++ controllers/api/userRoutes.js
@@ -1,6 +1,7 @@
 const router = require('express').Router();
 const { User } = require('../../models');
 
+//http://localhost:3001/user/
 router.post('/', async (req, res) => {
   console.log(req.body)
   try {
@@ -16,11 +17,11 @@ router.post('/', async (req, res) => {
     res.status(400).json(err);
   }
 });
-
+//http://localhost:3001/user/login
 router.post('/login', async (req, res) => {
   try {
     const userData = await User.findOne({ where: { email: req.body.email } });
-
+    console.log(req.body);
     if (!userData) {
       res
         .status(400)
@@ -40,7 +41,7 @@ router.post('/login', async (req, res) => {
     req.session.save(() => {
       req.session.user_id = userData.id;
       req.session.logged_in = true;
-      
+
       res.json({ user: userData, message: 'You are now logged in!' });
     });
 
@@ -49,6 +50,7 @@ router.post('/login', async (req, res) => {
   }
 });
 
+//http://localhost:3001/user/logout
 router.post('/logout', (req, res) => {
   if (req.session.logged_in) {
     req.session.destroy(() => {
