+++ models/Jar.js
@@ -1,35 +1,36 @@
 const { Model, DataTypes } = require('sequelize');
 const sequelize = require('../config/connection');
 
-class Jar extends Model {}
-Jar.init( { id: {
-      type: DataTypes.INTEGER,
-      allowNull: false,
-      primaryKey: true,
-      autoIncrement: true,
-    },
-    jarname:{
-      type: DataTypes.STRING,
-      allowNull:false
-    },
-    date_created: {
-      type: DataTypes.DATE,
-      allowNull: false,
-      defaultValue: DataTypes.NOW,
-    },
-    daily_earnings: {
-      type: DataTypes.DECIMAL,
-      allowNull: false,
-    },
-    user_id: {
-      type: DataTypes.INTEGER,
-      // foreign key
-      references: {
-        model: 'user',
-        key: 'id',
-      },
+class Jar extends Model { }
+Jar.init({
+  id: {
+    type: DataTypes.INTEGER,
+    allowNull: false,
+    primaryKey: true,
+    autoIncrement: true,
+  },
+  jarname: {
+    type: DataTypes.STRING,
+    allowNull: false
+  },
+  date_created: {
+    type: DataTypes.DATE,
+    allowNull: false,
+    defaultValue: DataTypes.NOW,
+  },
+  daily_earnings: {
+    type: DataTypes.DECIMAL,
+    allowNull: false,
+  },
+  user_id: {
+    type: DataTypes.INTEGER,
+    // foreign key
+    references: {
+      model: 'user',
+      key: 'id',
     },
   },
+},
   {
     sequelize,
     timestamps: false,
