+++ views/homepage.handlebars
@@ -107,6 +107,8 @@
         </div>
       </div>
     </div>
+
+    {{!-- GRADIENT EFFECT --}}
     <div
       class="absolute inset-x-0 top-[calc(100%-13rem)] -z-10 transform-gpu overflow-hidden blur-3xl sm:top-[calc(100%-30rem)]"
       aria-hidden="true">
