+++ views/login.handlebars
@@ -49,6 +49,13 @@
 
 {{!-- TAILWINDCSS CODE IS BELOW --}}
 <div class="flex min-h-full flex-col justify-center px-6 py-12 lg:px-8">
+  {{!-- GRADIENT EFFECT --}}
+  <div class="absolute inset-x-0 -top-40 -z-10 transform-gpu overflow-hidden blur-3xl sm:-top-80" aria-hidden="true">
+    <div
+      class="relative left-[calc(50%-11rem)] aspect-[1155/678] w-[36.125rem] -translate-x-1/2 rotate-[30deg] bg-gradient-to-tr from-[#ff80b5] to-[#9089fc] opacity-30 sm:left-[calc(50%-30rem)] sm:w-[72.1875rem]"
+      style="clip-path: polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)">
+    </div>
+  </div>
   {{!-- logo --}}
   <div class="sm:mx-auto sm:w-full sm:max-w-sm">
     <a href="/">
@@ -85,17 +92,26 @@
       {{!-- LOGIN BUTTON --}}
       <div class="form-group">
         <button type="submit"
-          class="flex w-full justify-center rounded-md bg-indigo-600 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600">Sign
+          class="flex w-full justify-center rounded-md bg-indigo-600 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-green-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600">Sign
           in</button>
       </div>
-
-      {{!-- SIGNUP REDIRECTION --}}
     </form>
+    {{!-- SIGNUP REDIRECTION --}}
     <p class="mt-10 text-center text-sm text-gray-500">
       Don't have an account?
-      <a href="/signup" class="font-semibold leading-6 text-indigo-600 hover:text-indigo-500">Sing up here.</a>
+      <a href="/signup" class="font-semibold leading-6 text-indigo-600 hover:text-green-500">Sing up here.</a>
     </p>
   </div>
 
-  <script src="./js/login.js"></script>
+  {{!-- GRADIENT EFFECT --}}
+  <div
+    class="absolute inset-x-0 top-[calc(100%-13rem)] -z-10 transform-gpu overflow-hidden blur-3xl sm:top-[calc(100%-30rem)]"
+    aria-hidden="true">
+    <div
+      class="relative left-[calc(50%+3rem)] aspect-[1155/678] w-[36.125rem] -translate-x-1/2 bg-gradient-to-tr from-[#ff80b5] to-[#9089fc] opacity-30 sm:left-[calc(50%+36rem)] sm:w-[72.1875rem]"
+      style="clip-path: polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)">
+    </div>
+  </div>
+
+  <script src="/js/login.js"></script>
 </div>
