+++ views/profile.handlebars
@@ -150,6 +150,7 @@
           </ul>
         </div>
 
+        {{!-- MONTHLY GOAL CARD --}}
         <div class="-mt-2 p-2 lg:mt-0 lg:w-full lg:max-w-md lg:flex-shrink-0">
           <div
             class="rounded-2xl bg-gray-50 py-10 text-center ring-1 ring-inset ring-gray-900/5 lg:flex lg:flex-col lg:justify-center lg:py-16">
